+++ src/ts/Ads/Utilities/SDKMetrics.ts
@@ -161,7 +161,10 @@ export enum MediationMetric {
     MediaCount = 'media_count',
     AuctionRequest = 'auction_request_time',
     AdCaching = 'ad_caching_time',
-    AuctionRequestStarted = 'auction_request_start'
+    AuctionRequestStarted = 'auction_request_start',
+    FillLatencyByPlacements = 'fill_latency_by_placements',
+    NofillLatencyByPlacements = 'nofill_latency_by_placements',
+    InitCompleteByPlacements = 'mediation_init_complete_by_placements'
 }
 
 export enum LoadV5 {
@@ -187,10 +190,7 @@ export enum AuctionV6 {
     FailedToParse = 'v6_failed_to_parse',
     AuctionIdMissing = 'v6_auction_id_missing',
     PlacementsMissing = 'v6_placements_missing',
-    MediaIdMissing = 'v6_media_id_missing',
-    TrackingMissing = 'v6_tracking_missing',
     TrackingIndicesOutOfBounds = 'v6_tracking_indices_out_of_bounds',
-    BannerPlacementNotRemoved = 'v6_banner_placement_not_removed',
     FailedCreatingAuctionResponse = 'v6_failed_creating_auction_response'
 }