+++ src/ts/Core/Models/ABGroup.ts
@@ -77,7 +77,6 @@ export const FakeReverseABTest = new ReverseABTest(16, 17);
 
 // Active AB Tests
 export const MediationCacheModeAllowedTest = new ReverseABTest(5);
-export const TemporaryCacheModeAllowedTest = new ABTest(14, 15);
 export const MabDisabledABTest = new ZyngaFilteredABTest(7);
 export const PrivacySDKTest = new ABTest(16);
 export const AuctionXHR = new DisabledABTest();